TRAVEL_PLANNER_WITH_ADK/pyproject.toml
[project]
name = "travel-planner-with-adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ag-ui-adk>=0.4.2",
    "fastapi>=0.123.10",
    "geopy>=2.4.1",
    "google-adk>=1.23.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
